Use web testcase & dynamic routes from the pimcore testing framework
