+++ tests/Integration/Tagging/Fixtures/get_asset_route.yaml
@@ -1,0 +1,3 @@
+get_asset:
+  path: /get-asset
+  controller: App\Controller\GetAssetController
