+++ tests/Integration/Tagging/Fixtures/get_document_route.yaml
@@ -1,0 +1,4 @@
+get_document:
+    path: /get-document
+    controller: App\Controller\GetDocumentController
+
