+++ tests/Integration/Tagging/Fixtures/get_object_route.yaml
@@ -1,0 +1,3 @@
+get_object:
+    path: /get-object
+    controller:  App\Controller\GetObjectController
