+++ tests/Integration/Tagging/TagAssetTest.php
@@ -1,17 +1,55 @@
 <?php declare(strict_types=1);
 
-namespace Integration;
+namespace Integration\Tagging;
 
 use Neusta\Pimcore\HttpCacheBundle\CacheActivator;
-use Neusta\Pimcore\HttpCacheBundle\Tests\Integration\ConfigurableWebTestcase;
 use Neusta\Pimcore\TestingFramework\Database\ResetDatabase;
 use Neusta\Pimcore\TestingFramework\Test\Attribute\ConfigureExtension;
+use Neusta\Pimcore\TestingFramework\Test\Attribute\ConfigureRoute;
+use Neusta\Pimcore\TestingFramework\Test\ConfigurableWebTestcase;
 use Pimcore\Cache\RuntimeCache;
-use Pimcore\Model\DataObject\TestDataObject;
+use Pimcore\Model\Asset;
 
-final class TagObjectTest extends ConfigurableWebTestcase
+#[ConfigureRoute(__DIR__ . '/Fixtures/get_asset_route.yaml')]
+final class TagAssetTest extends ConfigurableWebTestcase
 {
     use ResetDatabase;
+
+    /**
+     * @test
+     */
+    #[ConfigureExtension('neusta_pimcore_http_cache', [
+        'elements' => [
+            'assets' => true,
+            'documents' => false,
+            'objects' => false,
+        ],
+    ])]
+    public function response_is_tagged_with_expected_tags_when_asset_is_loaded(): void
+    {
+        $client = self::createClient();
+
+        $asset = new Asset();
+        $asset->setId(42);
+        $asset->setFilename('test-asset.txt');
+        $asset->setParentId(1);
+        $asset->setData('This is the content of the test asset.');
+        $asset->setMimetype('text/plain');
+        $asset->save();
+
+        // Clear the runtime cache, as it prevents the object from being loaded and thus tagged.
+        // Note: in reality, objects are created and loaded/used in separate requests.
+        RuntimeCache::clear();
+
+        $client->request('GET', '/get-asset?id=42');
+
+        $response = $client->getResponse();
+        self::assertSame('This is the content of the test asset.', $response->getContent());
+        self::assertSame(200, $response->getStatusCode());
+        self::assertTrue($response->headers->getCacheControlDirective('public'));
+        self::assertSame('3600', $response->headers->getCacheControlDirective('s-maxage'));
+        self::assertSame('a42', $response->headers->get('X-Cache-Tags'));
+    }
 
     /**
      * @test
@@ -23,62 +61,26 @@ final class TagObjectTest extends ConfigurableWebTestcase
             'objects' => true,
         ],
     ])]
-    public function response_is_tagged_with_expected_tags_when_object_is_loaded(): void
+    public function response_is_not_tagged_when_assets_is_not_enabled(): void
     {
         $client = self::createClient();
 
-        $object = new TestDataObject();
-        $object->setId(42);
-        $object->setKey('test_object');
-        $object->setContent('Test content');
-        $object->setPublished(true);
-        $object->setParentId(1);
-        $object->save();
+        $asset = new Asset();
+        $asset->setId(42);
+        $asset->setFilename('test-asset.txt');
+        $asset->setParentId(1);
+        $asset->setData('This is the content of the test asset.');
+        $asset->setMimetype('text/plain');
+        $asset->save();
 
         // Clear the runtime cache, as it prevents the object from being loaded and thus tagged.
         // Note: in reality, objects are created and loaded/used in separate requests.
         RuntimeCache::clear();
 
-        $client->request('GET', '/get-object?id=42');
+        $client->request('GET', '/get-asset?id=42');
 
         $response = $client->getResponse();
-        self::assertSame('Test content', $response->getContent());
-        self::assertSame(200, $response->getStatusCode());
-        self::assertTrue($response->headers->getCacheControlDirective('public'));
-        self::assertSame('3600', $response->headers->getCacheControlDirective('s-maxage'));
-        self::assertSame('o42', $response->headers->get('X-Cache-Tags'));
-    }
-
-    /**
-     * @test
-     */
-    #[ConfigureExtension('neusta_pimcore_http_cache', [
-        'elements' => [
-            'assets' => false,
-            'documents' => false,
-            'objects' => false,
-        ],
-    ])]
-    public function response_is_not_tagged_when_objects_is_not_enabled(): void
-    {
-        $client = self::createClient();
-
-        $object = new TestDataObject();
-        $object->setId(42);
-        $object->setKey('test_object');
-        $object->setContent('Test content');
-        $object->setPublished(true);
-        $object->setParentId(1);
-        $object->save();
-
-        // Clear the runtime cache, as it prevents the object from being loaded and thus tagged.
-        // Note: in reality, objects are created and loaded/used in separate requests.
-        RuntimeCache::clear();
-
-        $client->request('GET', '/get-object?id=42');
-
-        $response = $client->getResponse();
-        self::assertSame('Test content', $response->getContent());
+        self::assertSame('This is the content of the test asset.', $response->getContent());
         self::assertSame(200, $response->getStatusCode());
         self::assertTrue($response->headers->getCacheControlDirective('public'));
         self::assertSame('3600', $response->headers->getCacheControlDirective('s-maxage'));
@@ -90,9 +92,9 @@ final class TagObjectTest extends ConfigurableWebTestcase
      */
     #[ConfigureExtension('neusta_pimcore_http_cache', [
         'elements' => [
-            'assets' => false,
+            'assets' => true,
             'documents' => false,
-            'objects' => true,
+            'objects' => false,
         ],
     ])]
     public function response_is_not_tagged_when_caching_is_deactivated(): void
@@ -101,22 +103,22 @@ final class TagObjectTest extends ConfigurableWebTestcase
 
         static::getContainer()->get(CacheActivator::class)->deactivateCaching();
 
-        $object = new TestDataObject();
-        $object->setId(42);
-        $object->setKey('test_object');
-        $object->setContent('Test content');
-        $object->setPublished(true);
-        $object->setParentId(1);
-        $object->save();
+        $asset = new Asset();
+        $asset->setId(42);
+        $asset->setFilename('test-asset.txt');
+        $asset->setParentId(1);
+        $asset->setData('This is the content of the test asset.');
+        $asset->setMimetype('text/plain');
+        $asset->save();
 
         // Clear the runtime cache, as it prevents the object from being loaded and thus tagged.
         // Note: in reality, objects are created and loaded/used in separate requests.
         RuntimeCache::clear();
 
-        $client->request('GET', '/get-object?id=42');
+        $client->request('GET', '/get-asset?id=42');
 
         $response = $client->getResponse();
-        self::assertSame('Test content', $response->getContent());
+        self::assertSame('This is the content of the test asset.', $response->getContent());
         self::assertSame(200, $response->getStatusCode());
         self::assertTrue($response->headers->getCacheControlDirective('public'));
         self::assertSame('3600', $response->headers->getCacheControlDirective('s-maxage'));
