--- tests/app/config/routes.yaml
@@ -1,5 +1,0 @@
-controllers:
-  resource:
-    path: ../src/Controller/
-    namespace: App\Controller
-  type: attribute
