+++ tests/app/src/Controller/GetAssetController.php
@@ -7,12 +7,10 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\HttpKernel\Attribute\Cache;
-use Symfony\Component\Routing\Attribute\Route;
 
 #[AsController]
 final class GetAssetController
 {
-    #[Route(path: '/get-asset', name: 'get_asset')]
     #[Cache(smaxage: 3600, public: true)]
     public function __invoke(Request $request): Response
     {
