+++ tests/app/src/Controller/GetDocumentController.php
@@ -6,11 +6,9 @@ use Pimcore\Model\Document;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\Cache;
-use Symfony\Component\Routing\Attribute\Route;
 
 final class GetDocumentController
 {
-    #[Route(path: '/get-document', name: 'get_document')]
     #[Cache(smaxage: 3600, public: true)]
     public function __invoke(Request $request): Response
     {
