+++ tests/app/src/Controller/GetObjectController.php
@@ -1,5 +1,4 @@
-<?php
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 namespace App\Controller;
 
@@ -8,12 +7,10 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\HttpKernel\Attribute\Cache;
-use Symfony\Component\Routing\Attribute\Route;
 
 #[AsController]
 final class GetObjectController
 {
-    #[Route(path: '/get-object', name: 'get_object')]
     #[Cache(smaxage: 3600, public: true)]
     public function __invoke(Request $request): Response
     {
